--- src/extensions/components/InstallExtensionManifestData/index.ts
@@ -1,1 +1,0 @@
-export * from "./InstallExtensionManifestData";
