+++ src/extensions/views/InstallCustomExtension/index.ts
@@ -1,0 +1,1 @@
+export * from "./InstallCustomExtension";
